Add files via upload
EER Diagram of Greenspot dataset. 

Let’s discuss the relationships and cardinalities among the tables. The Product table is related to the Stock table with a one-to-many relationship, meaning each product can have many records in the stock table. This is appropriate because the stock of a product will always be updated to reflect how much stock is available at a given time.

Next, the relationship between the Product table and Purchase table is one-to-many, indicating that the store can make purchases of the same product more than once. Following that, the relationship between the Product table and the Salesdetails table shows a one-to-many relationship, meaning a product can be sold multiple times. The relationship between the Customer table and Sales table is also one-to-many, signifying that each customer can make more than one purchase.